separated dto classes
